Refactor helper functions
